requirements.txt
django >= 1.8, < 1.9
djangorestframework==3.3.1
